First commit: basic reminder utility
    This first incarnation is rough, but it does the job. The Album.__init__ is
repetitive, and needs to be refactored, if possible using optional subpatterns
in a single regex. The monthly list is also fragile, depending on a run on the
1st of the month. It may be simplest to leave the intervals to the user; this
fix might also be helped by daemonization, rather than depending on cron.
    On the email side, there are two issues: the need for a local SMTP server,
and lack of metadata in the message. The server may be replaced by the user's
email service of choice, although authentication is an unknown right now. I'm
sure it's in the docs. As for metadata, this can be easily fixed with a MIME
message tomorrow.